[Bernstein&Kenneth Garcia] Moved all logic from CompanyController to CompanyService

+++ src/main/java/com/tw/apistackbase/service/CompanyService.java
@@ -4,7 +4,8 @@ import com.tw.apistackbase.core.Company;
 import com.tw.apistackbase.repository.CompanyRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -14,13 +15,22 @@ public class CompanyService {
     private CompanyRepository companyRepository;
 
     public Iterable<Company> findAllByNameLike(String name, Integer page, Integer pageSize) {
-        Pageable pageable = PageRequest.of(page, pageSize);
-        return companyRepository.findAllByNameLike(name, pageable);
+        return companyRepository.findAllByNameLike(name, PageRequest.of(page, pageSize));
     }
 
     public Iterable<Company> findAll(Integer page, Integer pageSize) {
-        Pageable pageable = PageRequest.of(page, pageSize);
-        return companyRepository.findAll(pageable);
+        return companyRepository.findAll(PageRequest.of(page, pageSize));
+    }
+
+    public ResponseEntity<Company> updateCompanyByName(String name, Company updatedCompany) {
+        Company company = companyRepository.findOneByName(name);
+        if (company == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        company.setName(updatedCompany.getName());
+        company.setProfile(updatedCompany.getProfile());
+        company.setEmployees(updatedCompany.getEmployees());
+        return new ResponseEntity<>(company, HttpStatus.OK);
     }
 
     public Company findOneByName(String name) {
@@ -31,7 +41,12 @@ public class CompanyService {
         return companyRepository.save(company);
     }
 
-    public void deleteByName(String name) {
+    public ResponseEntity<Company> deleteCompanyByName(String name) {
+        Company company = companyRepository.findOneByName(name);
+        if (company == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         companyRepository.deleteByName(name);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 }
